refactor: move utils
move utils to nerdlets/shared/utils
