+++ nerdlets/mande-nerdlet/components/category-detail/ChartGrid.js
@@ -14,8 +14,8 @@ import {
   PieChart,
   FunnelChart,
 } from 'nr1'
-import { dateFormatter } from '../../../../utils/date-formatter'
-import { formatFacets } from '../../../../utils/query-formatter'
+import { dateFormatter } from '../../../shared/utils/date-formatter'
+import { formatFacets } from '../../../shared/utils/query-formatter'
 import { getHook } from '../../../shared/config/Hooks'
 
 const chartGrid = props => {
