+++ nerdlets/shared/utils/number-formatter.js
