+++ nerdlets/shared/utils/query-formatter.js
