+++ nerdlets/user-video-view/UserViewContainer.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Stack, StackItem, HeadingText } from 'nr1'
-import { formatSinceAndCompare } from '../../utils/query-formatter'
-import { dateFormatter } from '../../utils/date-formatter'
+import { formatSinceAndCompare } from '../shared/utils/query-formatter'
+import { dateFormatter } from '../shared/utils/date-formatter'
 import QosKpiGrid from '../shared/components/qos/QosKpiGrid'
 import videoConfig from '../shared/config/VideoConfig'
 import ViewTable from './components/view/ViewTable'
